refactor(notifications): extract RoutableToastNotification contract and DelegatesToToastNotification trait
The new absence notifications and SubscribableNotification re-declared
the same toArray/toWebPush/toFcm/toFcmData boilerplate. Hoist the
shared surface into a contract and the default channel translations
into a trait so notifications only need to provide
toToastNotification() (and optionally getRoute()).

- Add FluxErp\Contracts\RoutableToastNotification, extending
  HasToastNotification with getRoute, toArray, toWebPush, toFcm and
  toFcmData
- Add FluxErp\Traits\Notification\DelegatesToToastNotification, the
  default implementation that delegates every channel through
  toToastNotification()
- SubscribableNotification implements the new contract and uses the
  trait, dropping the duplicated method bodies
- All six AbsenceRequest notifications adopt the same contract and
  trait, return $this->absenceRequest->getUrl() from getRoute() and
  consume that in their accept action instead of inlining the URL
