+++ src/Notifications/AbsenceRequest/AbsenceRequestSubstituteAssignedNotification.php
@@ -2,25 +2,25 @@
 
 namespace FluxErp\Notifications\AbsenceRequest;
 
-use FluxErp\Contracts\HasToastNotification;
+use FluxErp\Contracts\RoutableToastNotification;
 use FluxErp\Models\AbsenceRequest;
 use FluxErp\Notifications\Notification;
 use FluxErp\Support\Notification\ToastNotification\NotificationAction;
 use FluxErp\Support\Notification\ToastNotification\ToastNotification;
 use FluxErp\Traits\Makeable;
+use FluxErp\Traits\Notification\DelegatesToToastNotification;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
-use NotificationChannels\WebPush\WebPushMessage;
 
-class AbsenceRequestSubstituteAssignedNotification extends Notification implements HasToastNotification, ShouldQueue
+class AbsenceRequestSubstituteAssignedNotification extends Notification implements RoutableToastNotification, ShouldQueue
 {
-    use Makeable, Queueable;
+    use DelegatesToToastNotification, Makeable, Queueable;
 
     public function __construct(protected AbsenceRequest $absenceRequest) {}
 
-    public function toArray(object $notifiable): array
+    public function getRoute(): ?string
     {
-        return $this->toToastNotification($notifiable)->toArray();
+        return $this->absenceRequest->getUrl();
     }
 
     public function toToastNotification(object $notifiable): ToastNotification
@@ -37,18 +37,7 @@ class AbsenceRequestSubstituteAssignedNotification extends Notification implemen
             ->accept(
                 NotificationAction::make()
                     ->label(__('View'))
-                    ->url($this->absenceRequest->getUrl())
+                    ->url($this->getRoute())
             );
-    }
-
-    public function toWebPush(object $notifiable): ?WebPushMessage
-    {
-        if (! method_exists($notifiable, 'pushSubscriptions')
-            || ! $notifiable->pushSubscriptions()->exists()
-        ) {
-            return null;
-        }
-
-        return $this->toToastNotification($notifiable)->toWebPush();
     }
 }
