+++ src/Notifications/AbsenceRequest/AbsenceRequestSubstituteUnassignedNotification.php
@@ -2,24 +2,24 @@
 
 namespace FluxErp\Notifications\AbsenceRequest;
 
-use FluxErp\Contracts\HasToastNotification;
+use FluxErp\Contracts\RoutableToastNotification;
 use FluxErp\Models\AbsenceRequest;
 use FluxErp\Notifications\Notification;
 use FluxErp\Support\Notification\ToastNotification\ToastNotification;
 use FluxErp\Traits\Makeable;
+use FluxErp\Traits\Notification\DelegatesToToastNotification;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
-use NotificationChannels\WebPush\WebPushMessage;
 
-class AbsenceRequestSubstituteUnassignedNotification extends Notification implements HasToastNotification, ShouldQueue
+class AbsenceRequestSubstituteUnassignedNotification extends Notification implements RoutableToastNotification, ShouldQueue
 {
-    use Makeable, Queueable;
+    use DelegatesToToastNotification, Makeable, Queueable;
 
     public function __construct(protected AbsenceRequest $absenceRequest) {}
 
-    public function toArray(object $notifiable): array
+    public function getRoute(): ?string
     {
-        return $this->toToastNotification($notifiable)->toArray();
+        return $this->absenceRequest->getUrl();
     }
 
     public function toToastNotification(object $notifiable): ToastNotification
@@ -29,16 +29,5 @@ class AbsenceRequestSubstituteUnassignedNotification extends Notification implem
             ->title(__('You are no longer substitute for :employee', [
                 'employee' => $this->absenceRequest->employee?->name ?? __('Unknown'),
             ]));
-    }
-
-    public function toWebPush(object $notifiable): ?WebPushMessage
-    {
-        if (! method_exists($notifiable, 'pushSubscriptions')
-            || ! $notifiable->pushSubscriptions()->exists()
-        ) {
-            return null;
-        }
-
-        return $this->toToastNotification($notifiable)->toWebPush();
     }
 }
